official/nlp/gru/requirements.txt
nltk
numpy
onnxruntime-gpu